# bff-openapi.yaml
openapi: 3.0.3
info:
  title: my-token BFF API
  version: "1.0.0"
  description: >
    The BFF's session-authenticated JSON surface (milestone-3
    `.chief/milestone-3/_contract/API.md`), consumed only by the SPA
    running in the owner's own browser. Every request authenticates over
    a signed, HttpOnly session cookie, resolved by
    `internal/transport/bff`'s own session middleware
    (`bff.RequireJSONSession`) — not by anything in this spec: no
    `security` scheme is declared here on purpose, for the exact same
    reason `openapi.yaml`'s own header comment gives for that file: this
    spec's job is request/response *shape* validation (`gin-middleware`),
    not auth. A second, independent file from `openapi.yaml` on purpose
    (`_contract/API.md` "Two specs, not one") — the public API
    (Bearer-authenticated, agent-facing) and this BFF surface
    (session-authenticated, owner-facing browser only) are two disjoint
    proof-of-identity mechanisms by construction (I2/I12), and a single
    shared spec file would blur that boundary even though the code
    enforces it.

    Deliberately absent from this surface: `POST /todos` aside, there is
    no key issuance or rotation endpoint here (no `POST /keys`, no
    rotate) — same reasoning `openapi.yaml`'s own `/keys/{id}` comment
    gives: a raw key can't be returned safely over an HTTP response, so
    issuance and rotation stay CLI-only regardless of which surface is
...

...
servers:
  - url: /api/bff
paths:
...
  /todos:
    get:
      operationId: listTodos
      summary: List every todo.
      description: >
        milestone-4: todos are a shared collection (`_rules/_contract/
        DATA_MODEL.md`/`GOAL.md`'s Ownership model decision) — I3 no
        longer applies to this domain, on either surface. Returns every
        todo, created_at descending, unpaginated — same shape as
        `_rules/_contract/API.md`'s `GET /api/v1/todos`, not just the
        session owner's own.
      responses:
        "200":
          description: Every todo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTodo
      summary: Create a todo, attributed to the caller.
      description: >
        `createdBy` is always the resolved session owner — never accepted
        from the body (I1, same rule as the public API, different auth
        mechanism) — but is attribution only, never access-scoping
        (GOAL.md). `status` always starts `open`. The `created` event
        this produces is never client-specifiable as a distinct write
        (I16) — it only ever happens as this endpoint's own side effect.
        `clientRequestId` makes this idempotent the same way
        `POST .../events` is (I19). Calls
        `internal/domain/todo.Service.CreateTodo` directly — the same
        service instance and method the public API's handler calls
        (`_rules/_standard/ARCHITECTURE.md`'s shared-service-layer rule).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: The created todo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTodo
      summary: Get a todo by id.
      description: >
        milestone-4: every authenticated actor may read every todo (I3 no
        longer applies to this domain) — an unknown id is `not_found`,
        there is no "wrong owner" case left to produce it.
      responses:
        "200":
          description: The todo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTodo
      summary: Rename a todo.
      description: >
        milestone-4: the only field this endpoint still writes is
        `title` — `status`/`assigneeId`/`priority`/`dueDate` changes
        route through `POST .../events` instead (`status_changed`/
        `assigned`/`field_changed`), the single write path (I15/I18)
        those fields' permission/audit requirements actually need.
        `done` is gone entirely (replaced by `status`) — sending it is a
        `validation_error` (`hint: "done"`), not silently accepted and
        ignored. Internally still funnels through the same single write
        path as every other mutation (a `field_changed` event), hence
        `clientRequestId` (I19).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoRequest"
      responses:
        "200":
          description: The updated todo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
...
    Todo:
      type: object
      required:
        [
          id,
          title,
          status,
          assigneeId,
          assigneeHandle,
          priority,
          dueDate,
          createdBy,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          description: >
            milestone-4: replaces `done`. Fixed enum, not an
            owner-editable table (GOAL.md).
          enum: [open, in_progress, done, closed]
        assigneeId:
          type: string
          nullable: true
          description: References a user (any role).
        assigneeHandle:
          type: string
          nullable: true
          description: >
            milestone-4 fix-round (handle-exposure): assigneeId's owning
            user's handle, for display — null exactly when assigneeId is
            null. my-task's own task list/detail views show a plain handle
            (`~/gits/my-task/src/server/modules/task/task.queries.ts`'s
            `assigneeHandle` LEFT JOIN, `t.assignee`/`detail.assignee` on
            the wire), never a bare id, and task-7's report found this
            repo's own Todo had no equivalent. Additive: assigneeId is
            unchanged, still what a caller writes back
            (CreateTodoRequest.assigneeId, CreateTodoEventRequest's `to`
            for an `assigned` event) — this is display-only, openapi.yaml's
            own Todo carries the identical field for the same reason
            (my-task's agent-facing REST shows it too, not just the owner
            UI).
        priority:
          type: string
          nullable: true
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
          description: >
            milestone-4: replaces `ownerId` — attribution only, never
            access-scoping (GOAL.md).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TodoList:
      type: object
      required: [todos]
      properties:
        todos:
          type: array
          items:
            $ref: "#/components/schemas/Todo"
...
    CreateTodoRequest:
      type: object
      additionalProperties: false
      description: >
        `additionalProperties: false` so a stray `done` (milestone-1/2/3's
        removed field) is a `validation_error`, not silently dropped
        (`_contract/API.md`). No `status` field, deliberately (mirrors
        the public API's own `CreateTodoRequest`) — every created todo
        starts `open` regardless of what a caller asks for.
      required: [title, clientRequestId]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        assigneeId:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
          nullable: true
        clientRequestId:
          type: string
          minLength: 1
          description: >
            I19's idempotency key, on the `created` event this endpoint's
            own side effect writes — required, not optional, since
            `todo_events.client_request_id` is `NOT NULL UNIQUE` at the
            schema level for every row including `created` ones.
    UpdateTodoRequest:
      type: object
      additionalProperties: false
      description: >
        milestone-4: `title` only — see `updateTodo`'s own description
        for why `status`/`assigneeId`/`priority`/`dueDate` aren't here.
        `additionalProperties: false` so a stray `done` is a
        `validation_error`, not silently dropped.
      required: [title, clientRequestId]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        clientRequestId:
          type: string
          minLength: 1
          description: >
            I19's idempotency key — this endpoint now funnels through
            the same single write path (I15) as every other mutation.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            hint:
              type: string
  responses:
    BadRequest:
      description: The request violates a request-shape or validation rule.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: >
        Missing, expired, or wrong-role session cookie — same shape
        regardless of which check failed (I5). A behavior change from
        this package's own HTML `GET /` view: this JSON surface always
        answers 401, never a redirect (a `fetch` call can't follow a
        redirect the way a browser navigation does — the SPA's own
        AuthGate-equivalent hook is what turns a 401 into a redirect,
        client-side).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >
        Unknown id, or an id that exists but isn't the caller's (I3 —
        absence, not permission).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"